-Add additional DbgEng missing types -Add various DebugAdvanced.Request() methods -Fix DebugControl output methods not being declared as cdecl

+++ ClrDebug/Native/Struct/DEBUG_EVENT/EXCEPTION_RECORD64.cs
@@ -10,8 +10,8 @@ namespace ClrDebug
     {
         public NTSTATUS ExceptionCode;
         public ExceptionFlags ExceptionFlags;
-        public IntPtr ExceptionRecord;
-        public IntPtr ExceptionAddress;
+        public ulong ExceptionRecord;
+        public ulong ExceptionAddress;
         public int NumberParameters;
         public int __unusedAlignment;
         public fixed long ExceptionInformation[15];
